#66 - feat: 계좌 선택 컴포넌트 분리
